#2 - font pretendard 추가 및 homescreen 생성
