+++ package.json
@@ -35,6 +35,7 @@
     "eslint": "^8.19.0",
     "jest": "^29.6.3",
     "prettier": "2.8.8",
+    "react-native-asset": "^2.1.1",
     "react-test-renderer": "19.1.0",
     "typescript": "^5.8.3"
   },
