+++ react-native.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+  project: {
+    ios: {},
+    android: {},
+  },
+  assets: ['./src/assets/fonts/'],
+};
+
